rename weight file
