+++ runtime/westend/src/weights/mod.rs
@@ -16,4 +16,4 @@
 
 /// A collection of weight modules used for pallets in the runtime.
 
-pub mod balances;
+pub mod pallet_balances;
